Move to switch
